backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
nltk==3.8.1